movies_admin/requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
backoff==2.2.1
bandit==1.7.5
certifi==2023.7.22
charset-normalizer==3.3.0
Django==3.2
django-cors-headers==4.3.0
django-debug-toolbar==3.4.0
django-extensions==3.2.3
django-split-settings==1.2.0
dnspython==2.4.2
email-validator==2.0.0.post2
flake8==4.0.0
gitdb==4.0.10
GitPython==3.1.34
idna==3.4
iniconfig==2.0.0
markdown-it-py==3.0.0
mccabe==0.6.1
mdurl==0.1.2
packaging==23.1
pbr==5.11.1
pluggy==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.8.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyflakes==2.4.0
Pygments==2.16.1
pytest==7.4.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rich==13.5.2
smmap==5.0.0
sqlparse==0.4.4
stevedore==5.1.0
typing_extensions==4.8.0
urllib3==2.0.6
uWSGI==2.0.22
